Laplacian Spectral Decomposition

Returns the eigenvalue spectrum together with eigenvectors of a Laplacian corresponding to a network. This involves computing the eigendecomposition of a (symmetric) matrix, so it is computationally intense and may take some time. The decomposition of the normalized Laplacian \(L = I - D^{-1}A\) takes is computed through the decomposition of its symmetric version \(L = D^{-\frac{1}{2}}AD^{-\frac{1}{2}}\). See the package vignette for details.

get_spectral_decomp(g, type = "Normalized Laplacian", verbose = FALSE)

Arguments

g

the network in the [igraph] format

type

the Laplacian type, default "Normalized Laplacian". At the moment this is the only available option. For other types of Laplacians one should get autonomously the eigendecomposition, e.g. using eigen. See the package vignette for an example.

verbose

whether warnings have to be printed or not

Value

lambdas the eigenvalues of the Laplacian

`u_L` the matrix of left eigenvectors (rows)

`u_R` the matrix of right eigenvectors (columns)

References

Bertagnolli, G., & De Domenico, M. (2021). Diffusion geometry of multiplex and interdependent systems. Physical Review E, 103(4), 042301. doi:10.1103/PhysRevE.103.042301 arXiv: 2006.13032

See also