Plot distance matrix

Plot a heatmap of the distance matrix using geom_tile.

# S3 method for diffudist
plot(
  DM,
  col_palette = viridis(n = 11),
  log_scale = FALSE,
  cex = 1,
  show_dendro = TRUE,
  title = ""
)

Arguments

DM

a distance matrix

col_palette

a colour palette, previously it was set to the `spectral` palette of brewer.pal but now it is, by default, to the color-blind firendly viridis.

log_scale

logical. Default FALSE

cex

numerical value by which text should be magnified (default font size in theme is 11)

show_dendro

If the dendrogram resulting from hclust should be shown. Default TRUE

title

Title of the plot passed to labs. No title by default.

Value

plot ggplot