Plot distance matrix

Plot a heatmap of the distance matrix using ggplot. Useful if used with multiplot ([source](http://stackoverflow.com/questions/6673162/reproducing-lattice-dendrogram-graph-with-ggplot2))

plot_distance_matrix(
  DM,
  col_palette = viridis::viridis(n = 11),
  log_scale = F,
  cex = 1,
  show_dendro = TRUE,
  title = NULL
)

Arguments

DM
a distance matrix

col_palette
a colour palette, previously it was set to the `spectral` palette of [RColorBrewer::brewer.pal] but now it is, by default, to the color-blind firendly [viridis::viridis].

log_scale
logical. Default FALSE

cex
numerical value by which text should be magnified (default font size in [ggplot2:theme] is 11)

show_dendro
If the dendrogram resulting from [stats::hclust] should be shown. Default TRUE

title
Title of the plot passed to [ggplot]. No title by default

Value

plot [ggplot]